requirements.txt
fastapi
uvicorn[standard]
pandas
requests
vaderSentiment
apscheduler
python-dotenv
pydantic
streamlit
